requirements.txt
google-genai==0.2.2
feedparser==6.0.11
PyYAML==6.0.2
python-dateutil==2.9.0.post0
